Split pack registry boundaries
