+++ package.json
@@ -86,6 +86,8 @@
     "harness:check:challenge-stage-correspondence": "node tools/harness/check-challenge-stage-correspondence.js",
     "harness:check:stage1-late-parity": "node tools/harness/check-stage1-late-clear-parity.js",
     "harness:check:stage2-carryover": "node tools/harness/check-stage2-carryover-parity.js",
+    "harness:check:gameplay-adapter-boundaries": "node tools/harness/check-gameplay-adapter-boundaries.js",
+    "harness:check:pack-registry-boundaries": "node tools/harness/check-pack-registry-boundaries.js",
     "harness:check:platinum-pack-boot": "node tools/harness/check-platinum-pack-boot.js",
     "harness:build:evidence-cycle-dashboard": "node tools/harness/build-evidence-cycle-dashboard.js",
     "harness:check:evidence-cycle-dashboard": "node tools/harness/check-evidence-cycle-dashboard.js",
