+++ src/js/05-player-flow.js
@@ -1,10 +1,6 @@
 // Aurora-specific player lifecycle, harness control, and manual movement helpers.
 
-function start(){
- if(typeof currentGamePackPlayable==='function'&&!currentGamePackPlayable()){
-  showToast('This pack is a shell preview only for now.');
-  return;
- }
+function startAuroraGameplay(){
  if(typeof clearRuntimeLoopFault==='function')clearRuntimeLoopFault();
  stopAttractLoop();
  try{document.activeElement?.blur?.()}catch{}
